[Refactoring] Componets -> components
