+++ gui/components/__init__.py
@@ -1,13 +1,10 @@
-# 순환참조 오류가 생길수 있으니 가장 기본적인 common을 import 해주어야함.
+# 순환참조 오류가 생길수 있으니 가장 기본적인 common을 먼저 import 해주어야함.
 
 # common 내용
 from .common.RoundedPanel import RoundedPanel
 from .common.SVGButton import SVGButton
-
-# Components 내용
-from .EditButton import EditButton
-from .CodeBox import CodeBox
-from .PromptInputPanel import PromptInputPanel
+from .common.EditButton import EditButton
+from .common.CodeBox import CodeBox
 
 # side 내용
 from .side.ConversationPanel import ConversationPanel
@@ -17,3 +14,4 @@ from .side.Setting import Settings
 # chat 내용
 from .chat.AIChatBox import AIChatBox
 from .chat.MyChatBox import MyChatBox
+from .chat.PromptInputPanel import PromptInputPanel
